Adds single and paired leaf dataset generation script
This changeset includes a new script which allows the creation of a dataset of
single leaf and paried leaf point clouds.

A new notebook is added which shows example usage of the new script methods and
the resulting dataset. Example data has been included locally to allow this
notebook to function within this repository. Note that for the notebook to
function additional dependencies of `kd3` will need to be installed on the
host. A future cleanup may be to add a requirements.txt to this repository.

Additional changes include addition of some __init__.py files and changes to
get modules to load in the notebooks directory. Another future cleanup would be
to standardize the approach for module loading.